requirements.txt
selenium==4.15.2
python-dotenv==1.0.0
webdriver-manager==4.0.1
jinja2==3.1.2
pytest==7.4.3
pytest-html==4.1.1
requests==2.31.0
Pillow==10.1.0
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.3.2
attrs==23.1.0
idna==3.4
markupsafe==2.1.3
packaging==23.2
